+++ src/atlannot/atlas/__init__.py
@@ -12,10 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Module for working with annotation atlases."""
-from .align import get_misalignment, unfurl_regions, warp
 
-__all__ = [
-    "get_misalignment",
-    "unfurl_regions",
-    "warp",
-]
