+++ src/atlannot/notebook/__init__.py
@@ -12,12 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Tools for working in a jupyter notebook."""
-from .util import create_legend_handles, image_grid, print_misalignments
-from .volume_viewer import VolumeViewer
 
-__all__ = [
-    "VolumeViewer",
-    "create_legend_handles",
-    "image_grid",
-    "print_misalignments",
-]
