+++ tests/test_notebook.py
@@ -16,7 +16,7 @@ import matplotlib.patches as patches
 import numpy as np
 import pytest
 
-from atlannot.notebook import create_legend_handles
+from atlannot.notebook.util import create_legend_handles
 from atlannot.region_meta import RegionMeta
